Moved sample test to same package as sample and removed module-info.java from test
