--- sample/src/test/java/module-info.java
@@ -1,10 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-open module com.microsoft.gctoolkit.sample.test {
-    requires com.microsoft.gctoolkit.sample;
-
-    requires java.logging;
-    requires org.junit.jupiter.api;
-    requires org.junit.jupiter.engine;
-}
